feat(pwa): add dialog process and modified the page name
