--- apps/pwa/src/app/client/roulette/page.tsx
@@ -1,12 +1,0 @@
-"use client";
-export default function page() {
-    return (
-        <main>
-            <div className='flex flex-col justify-center items-center'>
-                <div className='font-roboto text-3xl font-bold mb-4'>
-                    Welcome
-                </div>
-            </div>
-        </main>
-    );
-}
