+++ apps/pwa/src/components/DoroCard/game-card.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState } from "react";
 import {
     Dialog,
     DialogContent,
@@ -6,8 +7,8 @@ import {
     DialogHeader,
     DialogTitle,
     DialogTrigger,
+    DialogClose
 } from "@/components/ui/dialog"
-import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
@@ -19,6 +20,11 @@ import {
     CardHeader,
     CardTitle,
 } from "@/components/ui/card";
+import { Spinner } from "@/components/Spinner";
+import { delay } from "@/lib/utils";
+import {
+    PartyPopperIcon
+} from "lucide-react";
 
 type GameCardProps = React.ComponentProps<typeof Card> & {
     id: string,
@@ -30,9 +36,25 @@ type GameCardProps = React.ComponentProps<typeof Card> & {
     creator: string
 };
 
+enum status {
+    default = 0,
+    success = 1,
+    error = 2
+}
+
 export function GameCard({ className, ...props }: GameCardProps) {
+    const [isLoading, setIsLoading] = useState(false);
+    const [open, setOpen] = useState(false);
+    const [isFinished, setFinished] = useState(false)
+
+    useEffect(() => {
+        if (!open) {
+            setFinished(false)
+        }
+    }, [open])
+
     return (
-        <Dialog>
+        <Dialog open={open} onOpenChange={setOpen}>
             <DialogTrigger asChild>
                 <Card
                     className={cn(
@@ -60,44 +82,89 @@ export function GameCard({ className, ...props }: GameCardProps) {
                 </Card>
             </DialogTrigger>
             <DialogContent className="max-w-[250px]">
-                <DialogHeader>
-                    <DialogTitle>Join Game ?</DialogTitle>
-                </DialogHeader>
-                <div className="grid gap-4 py-4">
-                    <div className="grid grid-cols-4 items-center gap-4">
-                        <Label htmlFor="GAME" className="text-right">
-                            GAME
-                        </Label>
-                        <Label className="col-span-3">
-                            {props.title}
-                        </Label>
-                    </div>
-                    <div className="grid grid-cols-4 items-center gap-4">
-                        <Label htmlFor="Prize" className="text-right">
-                            Prize
-                        </Label>
-                        <Label className="col-span-3">
-                            {props.prize}
-                        </Label>
-                    </div>
-                    <div className="grid grid-cols-4 items-center gap-4">
-                        <Label htmlFor="fee" className="text-right">
-                            fee
-                        </Label>
-                        <Label className="col-span-3">
-                            {props.fee}
-                        </Label>
-                    </div>
-                </div>
-                <DialogFooter>
-                    <Button
-                        onClick={() => {
-                            console.log(`DORO ${props.id}`)
-                        }}>DORO
-                    </Button>
-                </DialogFooter>
-            </DialogContent>
-        </Dialog>
+                {open
+                    ? (<>
+                        {!isFinished ? (
+                            <div>
+                                <DialogHeader>
+                                    <DialogTitle>Join Game ?</DialogTitle>
+                                </DialogHeader>
+                                <div className="grid gap-4 py-4">
+                                    <div className="grid grid-cols-4 items-center gap-4">
+                                        <Label htmlFor="GAME" className="text-right">
+                                            GAME
+                                        </Label>
+                                        <Label className="col-span-3">
+                                            {props.title}
+                                        </Label>
+                                    </div>
+                                    <div className="grid grid-cols-4 items-center gap-4">
+                                        <Label htmlFor="Prize" className="text-right">
+                                            Prize
+                                        </Label>
+                                        <Label className="col-span-3">
+                                            {props.prize}
+                                        </Label>
+                                    </div>
+                                    <div className="grid grid-cols-4 items-center gap-4">
+                                        <Label htmlFor="fee" className="text-right">
+                                            fee
+                                        </Label>
+                                        <Label className="col-span-3">
+                                            {props.fee}
+                                        </Label>
+                                    </div>
+                                </div>
+                                <DialogFooter>
+                                    {
+                                        !isLoading ? (
+                                            <Button
+                                                onClick={() => {
+                                                    console.log(`DORO ${props.id}`)
+                                                    setIsLoading(true);
+
+                                                    // TODO: Disconnect (Mock Loading)
+                                                    delay(500).finally(() => {
+                                                        setIsLoading(false);
+                                                        setFinished(true)
+                                                    });
+                                                }}>DORO
+                                            </Button>
+                                        ) : (
+                                            <div className="flex w-full justify-center">
+                                                <Spinner />
+                                            </div>
+                                        )
+                                    }
+                                </DialogFooter>
+                            </div>
+                        ) : (
+                            <div>
+                                <DialogHeader>
+                                    <DialogTitle>
+                                        Good Luck
+                                    </DialogTitle>
+                                    <div className="grid gap-4 py-4 items-center justify-center">
+                                        <PartyPopperIcon />
+                                    </div>
+                                </DialogHeader>
+                                <DialogFooter>
+                                    <DialogClose asChild>
+                                        <Button
+                                            onClick={() => {
+                                                console.log(`GoodLuck ${props.id}`)
+                                            }}>OK
+                                        </Button>
+                                    </DialogClose>
+                                </DialogFooter>
+                            </div>
+                        )}
+                    </>)
+                    : (<>
+                        { }
+                    </>)}
+            </DialogContent >
+        </Dialog >
 
     );
 }
